refactor: UniversiHeader component in its own file
