+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from "react-router-dom";
 import { ThemeProvider } from "@/pages/Settings/GroupThemeColorPage/ThemeContext";
-import { Header } from "@/components/UniversiHeader";
+import UniversiHeader from "@/components/UniversiHeader";
 import useUmami from "./hooks/useUmami";
 
 import "./App.less";
@@ -11,7 +11,7 @@ export function App() {
     return (
         <ThemeProvider>
             <div id="App">
-                <Header />
+                <UniversiHeader />
 
                 <div id="app-main">
                     <Outlet />
