+++ src/components/UniversiHeader/index.ts
@@ -1,0 +1,2 @@
+export { UniversiHeader as default } from "./UniversiHeader";
+export * from "./UniversiHeader";
