+++ src/pages/Recovery/Recovery.tsx
@@ -1,7 +1,4 @@
-import { Header } from "@/components/UniversiHeader"
 import "./Recovery.css"
-import { transform } from "@babel/core"
-import { Translate } from "phosphor-react"
 import "../singin/signinForm.css"
 import {useState, useContext} from "react"
 import UniversimeApi from "@/services/UniversimeApi"
